requirements.txt
# ── Computer Vision ──────────────────────────────────────────────
ultralytics==8.2.18
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
scipy==1.13.0
supervision==0.27.0.post2

# ── ONNX Runtime ─────────────────────────────────────────────────
onnxruntime-gpu==1.18.0    # swap for onnxruntime on CPU-only machines

# ── Kafka ────────────────────────────────────────────────────────
aiokafka==0.10.0

# ── Database ─────────────────────────────────────────────────────
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.18.4
asyncpg==0.31.0

# ── API + Config ─────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.3.4     # NEW: needed for core/config.py

# ── LLM providers ────────────────────────────────────────────────
openai==1.30.1
google-generativeai==0.5.4
groq==0.9.0
requests==2.32.3              # used by OllamaClient
